# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a plant nursery point-of-sale system
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type EmployeeRole = typeof EmployeeRole[keyof typeof EmployeeRole];


export const EmployeeRole = {
  admin: 'admin',
  cashier: 'cashier',
} as const;

export interface Employee {
  id: number;
  nameAr: string;
  email: string;
  phone?: string | null;
  role: EmployeeRole;
  avatarUrl?: string | null;
  createdAt: string;
}

export type EmployeeInputRole = typeof EmployeeInputRole[keyof typeof EmployeeInputRole];


export const EmployeeInputRole = {
  admin: 'admin',
  cashier: 'cashier',
} as const;

export interface EmployeeInput {
  nameAr: string;
  email: string;
  phone?: string | null;
  role: EmployeeInputRole;
  password?: string;
  avatarUrl?: string | null;
}

export interface Category {
  id: number;
  nameAr: string;
  nameEn?: string | null;
  color?: string | null;
  icon?: string | null;
}

export interface CategoryInput {
  nameAr: string;
  nameEn?: string | null;
  color?: string | null;
  icon?: string | null;
}

export interface Supplier {
  id: number;
  name: string;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  createdAt: string;
}

export interface SupplierInput {
  name: string;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Product {
  id: number;
  nameAr: string;
  nameEn?: string | null;
  barcode: string;
  categoryId?: number | null;
  categoryNameAr?: string | null;
  supplierId?: number | null;
  purchasePrice: number;
  salePrice: number;
  quantity: number;
  lowStockThreshold: number;
  description?: string | null;
  imageUrl?: string | null;
  status: ProductStatus;
  createdAt: string;
}

export type ProductInputStatus = typeof ProductInputStatus[keyof typeof ProductInputStatus];


export const ProductInputStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ProductInput {
  nameAr: string;
  nameEn?: string | null;
  barcode?: string | null;
  categoryId?: number | null;
  supplierId?: number | null;
  purchasePrice: number;
  salePrice: number;
  quantity: number;
  lowStockThreshold?: number;
  description?: string | null;
  imageUrl?: string | null;
  status?: ProductInputStatus;
}

export interface Customer {
  id: number;
  name: string;
  phone?: string | null;
  avatarUrl?: string | null;
  balance: number;
  purchaseCount: number;
  totalSpent: number;
  createdAt: string;
}

export interface CustomerInput {
  name: string;
  phone?: string | null;
  avatarUrl?: string | null;
}

export interface InvoiceItem {
  id: number;
  productId: number;
  productNameAr: string;
  quantity: number;
  price: number;
  discount: number;
  notes?: string | null;
}

export interface InvoiceItemInput {
  productId: number;
  quantity: number;
  price: number;
  discount?: number;
  notes?: string | null;
}

export type InvoicePaymentMethod = typeof InvoicePaymentMethod[keyof typeof InvoicePaymentMethod];


export const InvoicePaymentMethod = {
  cash: 'cash',
  visa: 'visa',
  cliq: 'cliq',
  bank: 'bank',
  split: 'split',
} as const;

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  completed: 'completed',
  refunded: 'refunded',
} as const;

export interface Invoice {
  id: number;
  number: string;
  customerId?: number | null;
  customerName?: string | null;
  employeeId?: number | null;
  employeeName?: string | null;
  items: InvoiceItem[];
  subtotal: number;
  discount: number;
  tax: number;
  total: number;
  paymentMethod: InvoicePaymentMethod;
  status: InvoiceStatus;
  createdAt: string;
}

export type InvoiceInputPaymentMethod = typeof InvoiceInputPaymentMethod[keyof typeof InvoiceInputPaymentMethod];


export const InvoiceInputPaymentMethod = {
  cash: 'cash',
  visa: 'visa',
  cliq: 'cliq',
  bank: 'bank',
  split: 'split',
} as const;

export interface InvoiceInput {
  customerId?: number | null;
  employeeId?: number | null;
  items: InvoiceItemInput[];
  discount?: number;
  tax?: number;
  paymentMethod: InvoiceInputPaymentMethod;
}

export interface TopProduct {
  productId: number;
  nameAr: string;
  imageUrl?: string | null;
  totalSold: number;
  totalRevenue: number;
}

export interface LowStockProduct {
  productId: number;
  nameAr: string;
  quantity: number;
  lowStockThreshold: number;
}

export interface SalesPoint {
  date: string;
  sales: number;
  profit: number;
}

export interface RecentTransaction {
  id: number;
  number: string;
  customerName?: string | null;
  total: number;
  paymentMethod: string;
  createdAt: string;
}

export interface DashboardSummary {
  todaySales: number;
  todayProfit: number;
  monthlySales: number;
  invoiceCount: number;
  productCount: number;
  customerCount: number;
  supplierCount: number;
  employeeCount: number;
  topProducts: TopProduct[];
  lowStockProducts: LowStockProduct[];
  last7Days: SalesPoint[];
  recentTransactions: RecentTransaction[];
}

export type ListProductsParams = {
search?: string;
categoryId?: number;
};

export type ListCustomersParams = {
search?: string;
};

export type ListInvoicesParams = {
search?: string;
};