# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a plant nursery point-of-sale system
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeRole } from './employeeRole';

export interface Employee {
  id: number;
  nameAr: string;
  email: string;
  phone?: string | null;
  role: EmployeeRole;
  avatarUrl?: string | null;
  createdAt: string;
}